backend/pyproject.toml
[project]
name = "summiva"
version = "0.1.0"
description = "Enterprise-scale NLP system"
authors = [{ name = "Saikumar Bollam", email = "saikumar@arizona.edu" }]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "requests",
    "celery",
    "pymongo",
    "python-dotenv",
    "bson",
    "redis",
    "transformers",
    "scikit-learn",
    "spacy",
    "bertopic",
    "sentence-transformers",
    "elasticsearch",
    "faiss-cpu",
    "gunicorn"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "black",
    "isort",
    "ruff",
    "pre-commit",
    "mypy",
    "types-requests"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
